move reporter and test
